// src/app/models/driver.model.ts
export class driver {
    id?: string;
    userId?: string;
    userDescription?: string;
    firstName?: string;
    lastName?: string;
    phone?: string;
    email?: string;
    password?: string;
    idNumber?: string;
    dateOfBirth?: Date;
    licenceTypeId?: string;
    licenceTypeCode?: string;
    licenceExpiryDate?: Date;
    avatar?: string;
    statusId?: string;
    statusDescription?: string;
    review?: number;
    reviewCount?: number;
}